requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
faiss-cpu
requests
python-multipart
cryptography
